Make remaining suspicious implicit casts explicit
I ideally wanted to remove the need for these completely, but these are the
more difficult remaining ones which I don't have a good solution to.

For instance, (double)operationCount to do calculations is a common pattern,
and not a problem at the depths we see on the parseQueue. Converting through
NSNumber* could remove the casts, but I'm not convinced that's actually any
safer, and would lead to unnecessarily verbose code.

+++ Monal/Classes/MLDNSLookup.m
@@ -111,7 +111,7 @@ static NSMutableDictionary* _RRCache;
 
             struct timeval tv;
             tv.tv_sec  = (time_t)remainingTime;
-            tv.tv_usec = (int32_t)((remainingTime - tv.tv_sec) * 1000000);
+            tv.tv_usec = (int32_t)((remainingTime - (double)tv.tv_sec) * 1000000);
 
             int result = select(FD_SETSIZE, &set, NULL, NULL, &tv);
             DDLogVerbose(@"DNS select() returned %d", result);
